adding a styles at collections page and collectionsWall
