+++ src/App.styles.js
@@ -78,4 +78,27 @@ export const ImageOverlay = styled.div`
 		`}
 `;
 
-export const Container = styled.div``;
+export const UserLink = styled(StyledLink)`
+	padding: 5px;
+	display: flex;
+	align-items: center;
+	gap: 5px;
+`;
+
+export const Creator = styled.p`
+	color: white;
+	font-size: 18px;
+	font-weight: bold;
+`;
+
+export const SaveBtn = styled(Button)`
+	position: absolute;
+	bottom: 5%;
+	right: 5%;
+`;
+
+export const TotalLikes = styled(Creator)`
+	position: absolute;
+	bottom: 5%;
+	left: 5%;
+`;
