+++ src/Components/DisplayCarousel/DisplayCarousel.styles.js
@@ -11,7 +11,7 @@ export const View = styled.div`
 	width: 100%;
 	height: 100%;
 	display: ${(props) => props.display};
-	background-color: rgba(0, 0, 0, 0.1);
+	background-color: #767b82;
 	position: fixed;
 	top: 0;
 	left: 0;
