+++ src/Components/HighlightContents/HighlightContents.js
@@ -4,14 +4,15 @@ import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import { ReactComponent as Likes } from 'assets/likes.svg';
+import { ImageContainer } from './HighlightContents.styles';
 import {
-	ImageContainer,
+	Image,
+	ImageOverlay,
 	UserLink,
 	Creator,
 	SaveBtn,
 	TotalLikes,
-} from './HighlightContents.styles';
-import { Image, ImageOverlay } from 'App.styles';
+} from 'App.styles';
 
 export default class HighlightContents extends Component {
 	state = {
